requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pyodbc
azure-identity
